Move pinning mediator tests to Jest
